Abstracted Manager with FactoryManager
